feat: Add map image to projects and icon
Introduces a `mapImageUrl` field to the `Project` type and its mock data. This allows projects to display a map image alongside their main construction image.

A new 'map' icon has been added to the `Icon` component to support this feature. The `ProjectsView` component has been updated to conditionally display the map image when available and a toggle has been added to switch between the construction and map images.